Remove `background_color` and `clear_frame`
Resolves #100

I did find one display which was using `background_color` for stuff other than `clear_frame` - the epd5in16f was using it to set the border color. For now, I've replaced it with a `set_border_color` method which is only on that display. It should eventually be replaced by a proper method on `WaveshareDisplay`, but I don't have any hardware which supports border colors that I can test with.